requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-pytest==0.2.0
djangorestframework==3.16.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
